pyproject.toml
[tool.poetry]
package-mode = false
name = "quickstart"
version = "0.1.0"
description = ""
authors = ["thanhlt-1007 <le.tan.thanh@sun-asterisk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
ruff = "0.8.4"
django = "5.1.4"
djangorestframework = "3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
